Fix the env name validation
